feat: change animation control element
