+++ src/view/weatherMap/weatherMap.css
@@ -1,3 +1,8 @@
+:root {
+    --main-weather-map-color: rgb(212, 149, 0);
+}
+
+
 .weather-map {
     width: 100%;
     height: 800px;
@@ -142,14 +147,71 @@
 }
 
 .layers-buttons button.active {
-    background-color: rgb(212, 149, 0);
+    background-color: var(--main-weather-map-color);
 }
 
 .layers-buttons button.active:hover {
-    background-color: rgb(212, 149, 0);
+    background-color: var(--main-weather-map-color);
 }
 
-.wind-anim-checkbox-container {
+.wind-anim-toggler {
     display: flex;
+    width: 100%;
+    align-items: center;
+    cursor: pointer;
+    padding: 0.5rem 0;
+    font-size: 1.2rem;
+    color: #fff;
+    transition-duration: 300ms;
 }
 
+.wind-anim-toggler:hover {
+    text-shadow: 1px 1px 4px black;
+}
+
+.switch-btn {
+    display: inline-block;
+    width: 25px;
+    height: 12px;
+    box-sizing: content-box;
+    border: 1px solid white;
+    border-radius: 12px; 
+    background-color: rgba(0, 0, 0, 0.1);
+    z-index: 0;
+    margin: 0;
+    padding: 0;
+    cursor: pointer;
+    position: relative;
+    transition-duration: 300ms;
+}
+  
+.switch-btn::after {
+    content: "";
+    height: 12px;
+    width: 11px;
+    border-radius: 9px;
+    background-color: #fff;
+    left: 0px;
+    transition-duration: 300ms;
+    box-shadow: 0 0 4px 0 #999;
+    position: absolute;
+    z-index: 1;
+    -webkit-border-radius: 9px;
+    -moz-border-radius: 9px;
+    -ms-border-radius: 9px;
+    -o-border-radius: 9px;
+}
+
+.switch-on {
+    background-color: #fff;
+    box-shadow: inset 0 0 4px 0 #555;
+}
+
+.switch-on::after {
+    left: 14px;
+    background-color: var(--main-weather-map-color);
+}
+
+.wind-anim-label {
+    margin-left: 0.5rem;
+}
